+++ orix/vector/vector3d.py
@@ -591,10 +591,12 @@ class Vector3d(Object3d):
         axes_labels=None,
         vector_labels=None,
         hemisphere=None,
+        reproject=False,
         show_hemisphere_label=None,
         grid=None,
         grid_resolution=None,
         figure_kwargs=None,
+        reproject_scatter_kwargs=None,
         text_kwargs=None,
         return_figure=False,
         **kwargs,
@@ -619,6 +621,11 @@ class Vector3d(Object3d):
             otherwise adds to the current figure's hemispheres. Options
             are "upper", "lower", and "both", which plots two
             projections side by side.
+        reproject : bool, optional
+            Whether to reproject vectors onto the chosen hemisphere.
+            Reprojection is achieved by reflection of the vectors
+            located on the opposite hemisphere in the projection plane.
+            Ignored if `hemisphere` is "both". Default is False.
         show_hemisphere_label : bool, optional
             Whether to show hemisphere labels "upper" or "lower".
             Default is True if `hemisphere` is "both", otherwise False.
@@ -633,15 +640,23 @@ class Vector3d(Object3d):
         figure_kwargs : dict, optional
             Dictionary of keyword arguments passed to
             :func:`matplotlib.pyplot.subplots`.
+        reproject_scatter_kwargs: dict, optional
+            Dictionary of keyword arguments for the reprojected scatter
+            points which is passed to
+            :meth:`~orix.plot.StereographicPlot.scatter`, which passes
+            these on to :meth:`matplotlib.axes.Axes.scatter`. The
+            default marker style for reprojected vectors is "+". Values
+            used for vector(s) on the visible hemisphere are used unless
+            another value is passed here.
         text_kwargs : dict, optional
             Dictionary of keyword arguments passed to
-            :func:`~orix.plot.StereographicPlot.text`, which passes
+            :meth:`~orix.plot.StereographicPlot.text`, which passes
             these on to :meth:`matplotlib.axes.Axes.text`.
         return_figure : bool, optional
             Whether to return the figure (default is False).
         kwargs : dict, optional
             Keyword arguments passed to
-            :func:`~orix.plot.StereographicPlot.scatter`, which passes
+            :meth:`~orix.plot.StereographicPlot.scatter`, which passes
             these on to :meth:`matplotlib.axes.Axes.scatter`.
 
         Returns
@@ -661,6 +676,13 @@ class Vector3d(Object3d):
         --------
         orix.plot.StereographicPlot
         """
+
+        from orix.plot.stereographic_plot import _is_visible
+
+        if hemisphere is None or hemisphere.lower() != "both":
+            both_hemispheres = False
+        else:
+            both_hemispheres = True
         (
             fig,
             axes,
@@ -681,11 +703,28 @@ class Vector3d(Object3d):
             text_kwargs=text_kwargs,
             axes_labels=axes_labels,
         )
-
+        if reproject and not both_hemispheres:
+            # setup reproject scatter plotting args
+            if reproject_scatter_kwargs is None:
+                reproject_scatter_kwargs = {}
+            reproject_scatter_kwargs.setdefault("marker", "+")
+            # unless otherwise defined, copy normal scatter kwargs
+            for k, v in kwargs.items():
+                if k not in reproject_scatter_kwargs.keys():
+                    reproject_scatter_kwargs[k] = v
+            # multiplicative factor to project z-data to other hemisphere
+            factor = (1, 1, -1)
         # Use methods of the StereographicPlot class
         for i, ax in enumerate(axes):  # Assumes a maximum of two axes
             ax.hemisphere = hemisphere[i]
             ax.scatter(self, **kwargs)
+            # only reproject if both hemispheres are not shown
+            if reproject and not both_hemispheres:
+                visible = _is_visible(self.polar, ax.pole)
+                ax.scatter(
+                    self.__class__(self[~visible].data * factor),
+                    **reproject_scatter_kwargs,
+                )
             ax.stereographic_grid(grid[i], grid_resolution[0], grid_resolution[1])
             ax._stereographic_grid = grid[i]
             ax.set_labels(*axes_labels)